Move brotli4j into test sources
