+++ http/http-advanced-reactive/src/test/java/io/quarkus/ts/http/advanced/reactive/DevModeBrotli4JHttpIT.java
@@ -1,21 +1,22 @@
 package io.quarkus.ts.http.advanced.reactive;
 
+import io.quarkus.test.bootstrap.DevModeQuarkusService;
+import io.quarkus.test.scenarios.QuarkusScenario;
+import io.quarkus.test.services.DevModeQuarkusApplication;
+import io.quarkus.ts.http.advanced.reactive.brotli4j.Brotli4JHttpServerConfig;
+import io.quarkus.ts.http.advanced.reactive.brotli4j.Brotli4JResource;
+import io.quarkus.ts.http.advanced.reactive.brotli4j.Brotli4JRestMock;
+import io.restassured.response.Response;
+import jakarta.ws.rs.core.HttpHeaders;
+import jakarta.ws.rs.core.MediaType;
+import org.junit.jupiter.api.Test;
+
 import static io.quarkus.ts.http.advanced.reactive.Brotli4JHttpIT.CONTENT_LENGTH_DEFAULT_TEXT_PLAIN;
-import static io.quarkus.ts.http.advanced.reactive.Brotli4JResource.DEFAULT_TEXT_PLAIN;
+import static io.quarkus.ts.http.advanced.reactive.brotli4j.Brotli4JResource.DEFAULT_TEXT_PLAIN;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.nullValue;
-
-import jakarta.ws.rs.core.HttpHeaders;
-import jakarta.ws.rs.core.MediaType;
-
-import org.junit.jupiter.api.Test;
-
-import io.quarkus.test.bootstrap.DevModeQuarkusService;
-import io.quarkus.test.scenarios.QuarkusScenario;
-import io.quarkus.test.services.DevModeQuarkusApplication;
-import io.restassured.response.Response;
 
 @QuarkusScenario
 public class DevModeBrotli4JHttpIT {
