+++ http/http-advanced-reactive/src/test/java/io/quarkus/ts/http/advanced/reactive/brotli4j/Brotli4JRestMock.java
@@ -1,4 +1,4 @@
-package io.quarkus.ts.http.advanced.reactive;
+package io.quarkus.ts.http.advanced.reactive.brotli4j;
 
 import java.io.IOException;
 import java.io.InputStream;
